Epic reloading :sunglasses:

+++ modules/events.js
@@ -1,12 +1,78 @@
 const Discord = require("discord.js");
+const path = require("path");
+require("../types.js");
 
+let lastAttemptedLogins = []
+
+/**
+ * @param {PassthroughType} passthrough
+ */
 module.exports = function(passthrough) {
-	let { client, config, utils, db, commands, reloadEvent } = passthrough;
+	let { client, config, commands, reloadEvent, reloader, reactionMenus, queueManager } = passthrough;
 	let stdin = process.stdin;
 	let prefixes = [];
 	let statusPrefix = "&";
+	let starting = true;
+	if (client.readyAt != null) starting = false;
 
-	reloadEvent.once(__filename, () => {
+	let utils = require("./utilities.js")(passthrough)
+	reloader.useSync("./modules/utilities.js", utils)
+
+	utils.addTemporaryListener(client, "message", path.basename(__filename), manageMessage)
+	if (client.readyAt != null) manageReady()
+	else utils.addTemporaryListener(client, "ready", path.basename(__filename), manageReady)
+	utils.addTemporaryListener(client, "messageReactionAdd", path.basename(__filename), reactionEvent);
+
+	utils.addTemporaryListener(client, "messageUpdate", path.basename(__filename), (oldMessage, data) => {
+		if (data.constructor.name == "Message") manageMessage(data);
+		else if (data.content) {
+			let channel = client.channels.get(data.channel_id);
+			let message = new Discord.Message(channel, data, client);
+			manageMessage(message);
+		}
+	})
+
+	utils.addTemporaryListener(client, "disconnect", path.basename(__filename), (reason) => {
+		if (reason) console.log(`Disconnected with ${reason.code} at ${reason.path}.`);
+		if (lastAttemptedLogins.length) console.log(`Previous disconnection was ${Math.floor(Date.now()-lastAttemptedLogins.slice(-1)[0]/1000)} seconds ago.`)
+		lastAttemptedLogins.push(Date.now())
+		new Promise(resolve => {
+			if (lastAttemptedLogins.length >= 3) {
+				let oldest = lastAttemptedLogins.shift()
+				let timePassed = Date.now()-oldest
+				let timeout = 30000
+				if (timePassed < timeout) return setTimeout(() => resolve(), timeout - timePassed)
+			}
+			return resolve()
+		}).then(() => {
+			client.login(config.bot_token)
+		})
+	})
+
+	utils.addTemporaryListener(client, "error", path.basename(__filename), reason => {
+		if (reason) console.error(reason);
+	})
+
+	utils.addTemporaryListener(process, "unhandledRejection", path.basename(__filename), reason => {
+		if (reason && reason.code) {
+			if (reason.code == 10008) return;
+			if (reason.code == 50013) return;
+		}
+		if (reason) console.error(reason);
+		else console.log("There was an error but no reason");
+	})
+
+	utils.addTemporaryListener(client, "guildMemberUpdate", path.basename(__filename), async (oldMember, newMember) => {
+		if (newMember.guild.id != "475599038536744960") return;
+		if (!oldMember.roles.get("475599593879371796") && newMember.roles.get("475599593879371796")) {
+			let row = await utils.sql.get("SELECT * FROM Premium WHERE userID =?", newMember.id);
+			if (!row) await utils.sql.all("INSERT INTO Premium (userID, state) VALUES (?, ?)", [newMember.id, 1]);
+			else return;
+		}
+		else return;
+	})
+
+	/*reloadEvent.once(__filename, () => {
 		client.removeListener("message", manageMessage);
 		client.removeListener("messageUpdate", manageEdit);
 		client.removeListener("disconnect", manageDisconnect);
@@ -22,12 +88,7 @@ module.exports = function(passthrough) {
 	client.on("voiceStateUpdate", manageVoiceStateUpdate);
 	client.on("guildMemberUpdate", manageMemberUpdate);
 	process.on("unhandledRejection", manageRejection);
-	stdin.on("data", manageStdin);
-
-	async function manageStdin(input) {
-		input = input.toString();
-		try { console.log(await utils.stringify(eval(input))); } catch (e) { console.log(e.stack); }
-	}
+	stdin.on("data", manageStdin);*/
 
 	/**
 	 * @param {Discord.Message} msg
@@ -94,6 +155,7 @@ module.exports = function(passthrough) {
 				try {
 					require("request-promise")(`http://ask.pannous.com/api?input=${encodeURIComponent(chat)}`).then(async res => {
 						let data = JSON.parse(res);
+						console.log(data)
 						if (!data.sp("output.0.actions")) return msg.channel.send("Terribly sorry but my Ai isn't working as of recently (◕︵◕)\nHopefully, the issue gets resolved soon. Until then, why not try some of my other features?");
 						let text = data.output[0].actions.say.text.replace(/Jeannie/gi, client.user.username).replace(/Master/gi, msg.member ? msg.member.displayName : msg.author.username).replace(/Pannous/gi, owner.username);
 						if (text.length >= 2000) text = text.slice(0, 1999)+"…";
@@ -107,56 +169,30 @@ module.exports = function(passthrough) {
 		}
 	}
 
-	function manageEdit(oldMessage, data) {
-		if (data.constructor.name == "Message") manageMessage(data);
-		else if (data.content) {
-			let channel = client.channels.get(data.channel_id);
-			let message = new Discord.Message(channel, data, client);
-			manageMessage(message);
-		}
-	}
-
-	function manageReady() {
-		console.log(`Successfully logged in as ${client.user.username}`);
-		process.title = client.user.username;
-		utils.sql.all("SELECT * FROM AccountPrefixes WHERE userID = ?", [client.user.id]).then(result => {
+	async function manageReady() {
+		await utils.sql.all("SELECT * FROM AccountPrefixes WHERE userID = ?", [client.user.id]).then(result => {
 			prefixes = result.map(r => r.prefix);
 			statusPrefix = result.find(r => r.status).prefix;
 			console.log("Loaded "+prefixes.length+" prefixes");
+		});
+		if (starting) {
+			console.log(`Successfully logged in as ${client.user.username}`);
+			process.title = client.user.username;
+			utils.sql.all("SELECT * FROM RestartNotify WHERE botID = ?", [client.user.id]).then(result => {
+				result.forEach(row => {
+					let channel = client.channels.get(row.channelID);
+					if (!channel) {
+						let user = client.users.get(row.mentionID);
+						if (!user) console.log(`Could not notify ${row.mentionID}`);
+						else user.send("Restarted! Uptime: "+process.uptime().humanize("sec"));
+					} else channel.send("<@"+row.mentionID+"> Restarted! Uptime: "+process.uptime().humanize("sec"));
+				});
+				utils.sql.all("DELETE FROM RestartNotify WHERE botID = ?", [client.user.id]);
+			});
 			update();
 			client.setInterval(update, 300000);
-		});
-		utils.sql.all("SELECT * FROM RestartNotify WHERE botID = ?", [client.user.id]).then(result => {
-			result.forEach(row => {
-				let channel = client.channels.get(row.channelID);
-				if (!channel) {
-					let user = client.users.get(row.mentionID);
-					if (!user) console.log(`Could not notify ${row.mentionID}`);
-					else user.send("Restarted! Uptime: "+process.uptime().humanize("sec"));
-				} else channel.send("<@"+row.mentionID+"> Restarted! Uptime: "+process.uptime().humanize("sec"));
-			});
-			utils.sql.all("DELETE FROM RestartNotify WHERE botID = ?", [client.user.id]);
-		});
-	}
-
-	function manageDisconnect(reason) {
-		if (reason) console.log(`Disconnected with ${reason.code} at ${reason.path}\n\nReconnecting in 6sec`);
-		setTimeout(() => client.login(config.bot_token), 6000);
-	}
-
-	function manageError(reason) {
-		if (reason) console.error(reason);
-	}
-
-	function manageRejection(reason) {
-		if (reason && reason.code) {
-			if (reason.code == 10008) return;
-			if (reason.code == 50013) return;
 		}
-		if (reason) console.error(reason);
-		else console.log("There was an error but no reason");
 	}
-	utils.manageError = manageRejection;
 
 	/**
 	 * @param {Discord.GuildMember} oldMember
@@ -166,23 +202,107 @@ module.exports = function(passthrough) {
 		if (newMember.id == client.user.id) return;
 		let channel = oldMember.voiceChannel || newMember.voiceChannel;
 		if (!channel || !channel.guild) return;
-		let queue = utils.queueStorage.storage.get(channel.guild.id);
+		let queue = queueManager.storage.get(channel.guild.id);
 		if (!queue) return;
 		queue.voiceStateUpdate(oldMember, newMember);
 	}
 
 	/**
-	 * @param {Discord.GuildMember} oldMember
-	 * @param {Discord.GuildMember} newMember
+	 * Handles reactions as actions for the Discord.Client to perform
+	 * @param {Array} actions An Array of Objects of actions
 	 */
-	async function manageMemberUpdate(oldMember, newMember) {
-		if (newMember.guild.id != "475599038536744960") return;
-		if (!oldMember.roles.get("475599593879371796") && newMember.roles.get("475599593879371796")) {
-			let row = await utils.sql.get("SELECT * FROM Premium WHERE userID =?", newMember.id);
-			if (!row) await utils.sql.all("INSERT INTO Premium (userID, state) VALUES (?, ?)", [newMember.id, 1]);
-			else return;
+	Discord.Message.prototype.reactionMenu = function(actions) {
+		let message = this;
+		return new ReactionMenu(message, actions);
+	}
+
+	/**
+	 * Class that initiates a Discord.Message reaction menu
+	 */
+	class ReactionMenu {
+		/**
+		 * Create a new ReactionMenu
+		 * @param {Discord.Message} message
+		 * @param {Array} actions
+		 */
+		constructor(message, actions) {
+			this.message = message;
+			this.actions = actions;
+			reactionMenus[this.message.id] = this;
+			this.promise = this.react();
 		}
-		else return;
+		async react() {
+			for (let a of this.actions) {
+				a.messageReaction = await this.message.react(a.emoji).catch(new Function());
+			}
+		}
+		destroy(remove) {
+			delete reactionMenus[this.message.id];
+			if (remove) {
+				if (this.message.channel.type == "text") {
+					this.message.clearReactions().catch(new Function());
+				} else if (this.message.channel.type == "dm") {
+					this.actions.forEach(a => {
+						if (a.messageReaction) a.messageReaction.remove().catch(new Function());
+					});
+				}
+			}
+		}
+	}
+
+	function reactionEvent(messageReaction, user) {
+		let id = messageReaction.messageID;
+		let emoji = messageReaction.emoji;
+		if (user.id == client.user.id) return;
+		let menu = reactionMenus[id];
+		if (!menu) return;
+		let msg = menu.message;
+		let action = menu.actions.find(a => a.emoji == emoji || (a.emoji.name == emoji.name && a.emoji.id == emoji.id));
+		if (!action) return;
+		if ((action.allowedUsers && !action.allowedUsers.includes(user.id)) || (action.deniedUsers && action.deniedUsers.includes(user.id))) {
+			if (action.remove == "user") messageReaction.remove(user);
+			return;
+		}
+		switch (action.actionType) {
+		case "reply":
+			msg.channel.send(user.mention+" "+action.actionData);
+			break;
+		case "edit":
+			msg.edit(action.actionData);
+			break;
+		case "js":
+			action.actionData(msg, emoji, user, messageReaction, reactionMenus);
+			break;
+		}
+		switch (action.ignore) {
+		case "that":
+			menu.actions.find(a => a.emoji == emoji).actionType = "none";
+			break;
+		case "thatTotal":
+			menu.actions = menu.actions.filter(a => a.emoji != emoji);
+			break;
+		case "all":
+			menu.actions.forEach(a => a.actionType = "none");
+			break;
+		case "total":
+			menu.destroy(true);
+			break;
+		}
+		switch (action.remove) {
+		case "user":
+			messageReaction.remove(user);
+			break;
+		case "bot":
+			messageReaction.remove();
+			break;
+		case "all":
+			msg.clearReactions();
+			break;
+		case "message":
+			menu.destroy(true);
+			msg.delete();
+			break;
+		}
 	}
 
 	const presences = {
@@ -217,7 +337,7 @@ module.exports = function(passthrough) {
 		]
 	};
 
-	const update = () => {
+	function update() {
 		let now = new Date().toJSON();
 		let choice;
 		if (now.includes("-10-09")) choice = [["happy age++, Cadence <3", "STREAMING"]]
@@ -228,6 +348,7 @@ module.exports = function(passthrough) {
 		else if (now.includes("-12-")) choice = presences.christmas;
 		else if (now.includes("-01-01")) choice = presences.newYears;
 		else if (now.includes("-01-16")) choice = [["at my owner's bday party", "PLAYING"]];
+		else if (now.includes("-07-04")) choice = [["with fireworks 💥", "PLAYING"]];
 		else choice = presences.yearly;
 		const [name, type] = choice[Math.floor(Math.random() * choice.length)];
 		client.user.setActivity(`${name} | ${statusPrefix}help`, { type, url: 'https://www.twitch.tv/papiophidian/' });
